Change goBack() to take an optional count or hash argument.
goBack( to:object )

Forces jQTouch to go back to a certain page in the history. You can
pass a specific page ID, a number of pages to go back, or nothing to
go back one page. If the specified page can not be found, jQTouch will
go back one page by default. Example:

  jQT.goBack(2); // Goes back 2 pages in history
  jQT.goBack('#home'); // Goes to #home as long as it's in history
  jQT.goBack(); // Goes back 1 page in history